+++ src/pages/Register.jsx
@@ -1,14 +1,13 @@
 import React from 'react'
 import Header from '../Components/Layout/Header'
-import ContainerForm from '../Components/RegisterForm/Container'
-import Footer from '../Components/Layout/Footer'
+import RegisterForm from '../Components/RegisterForm/RegisterForm'
 
 const Register = ({action}) => {
   return (
     <div>
         <Header/>
-        <ContainerForm action={action}/>
-        <Footer/>
+        {/* <ContainerForm action={action}/> */}
+        <RegisterForm/>
     </div>
   )
 }
